requirements.txt
echo "Django==5.2.1
djangorestframework==3.15.0
django-filter==24.1
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0
django-extensions==3.2.3
Pillow==10.4.0
openpyxl==3.1.2
pytest==8.0.2
pytest-django==4.7.0
six==1.16.0
redis==5.0.1
asgiref==3.7.2
python-dateutil==2.8.2
pytz==2023.3" > requirements.txt